Add check if canonical id is already in the system

+++ push/sender/src/main/java/org/jboss/aerogear/unifiedpush/message/sender/GCMPushNotificationSender.java
@@ -171,17 +171,25 @@ public class GCMPushNotificationSender implements PushNotificationSender {
             //check if current index of result has canonical id
             String canonicalRegId = result.getCanonicalRegistrationId();
             if (canonicalRegId != null) {
-                // same device has more than one registration id: update it
-              
-                // find the installation and change its token to new canonical id
-                Installation installation = clientInstallationService.findInstallationForVariantByDeviceToken(variantID,registrationIDs.get(i));
-                installation.setDeviceToken(canonicalRegId);
-              
-                //update installation with the new token
-                logger.info(String.format("Based on returned canonical id from GCM, updating Android installations with registration id [%s] with new token [%s] ", registrationIDs.get(i), canonicalRegId));
-                clientInstallationService.updateInstallation(installation);
-              
-                //if the result contains a canonical id, there is no need to remove the current inactive token from data store
+                // same device has more than one registration id: update it, if needed!
+                // let's see if the canonical id is already in our system:
+                Installation installation = clientInstallationService.findInstallationForVariantByDeviceToken(variantID, canonicalRegId);
+
+                if (installation != null) {
+                    // ok, there is already a device, with newest/latest registration ID (aka canonical id)
+                    // It is time to remove the old reg id, to avoid duplicated messages in the future!
+                    inactiveTokens.add(registrationIDs.get(i));
+
+                } else {
+                    // since there is no registered device with newest/latest registration ID (aka canonical id),
+                    // this means the new token/regId was never stored on the server. Let's update the device and change its token to new canonical id:
+                    installation = clientInstallationService.findInstallationForVariantByDeviceToken(variantID,registrationIDs.get(i));
+                    installation.setDeviceToken(canonicalRegId);
+
+                    //update installation with the new token
+                    logger.info(String.format("Based on returned canonical id from GCM, updating Android installations with registration id [%s] with new token [%s] ", registrationIDs.get(i), canonicalRegId));
+                    clientInstallationService.updateInstallation(installation);
+                }
                 
             }
             else
@@ -199,8 +207,10 @@ public class GCMPushNotificationSender implements PushNotificationSender {
             }
         }
         
-        // trigger asynchronous deletion:
-        logger.info(String.format("Based on GCM error response codes, deleting %d invalid Android installations", inactiveTokens.size()));
-        clientInstallationService.removeInstallationsForVariantByDeviceTokens(variantID, inactiveTokens);
+        if (! inactiveTokens.isEmpty()) {
+            // trigger asynchronous deletion:
+            logger.info(String.format("Based on GCM response data and error codes, deleting %d invalid or duplicated Android installations", inactiveTokens.size()));
+            clientInstallationService.removeInstallationsForVariantByDeviceTokens(variantID, inactiveTokens);
+        }
     }
 }
